% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsplines.R
\name{generate_response_smooth}
\alias{generate_response_smooth}
\title{Generates simulated response for multilevel splines -- test function #2}
\format{list(x_list = x_list, y_list = y_list, e_list = e_list, true_mu = mu, z = z)
\describe{
This function is supposed to be combined with the other generation function.. but later.
  \item{x_list}{the length-J list of design matrices. The nrow of each element is between N_s and N_m}
  \item{y_list}{the length-J list of response vectors. The length of each element is between N_s and N_m.}
  \item{e_list}{the length-J list of error vectors. The length of each element is between N_s and N_m.}
  \item{true_mu}{the true mu vector of length J}
  \item{z}{the grid vector of length J}
}}
\usage{
generate_response_smooth(J, mod, e_sigma = 1, x_sigma = 1, z_sigma = 0.5,
  N_s, N_m = 200)
}
\arguments{
\item{J}{number of 'data' intervals}

\item{mod}{underlying model; either `lm` or `glm`}

\item{e_sigma}{error variance - around the mean function; data level.}

\item{x_sigma}{design matrix sigma}

\item{z_sigma}{error variance around my surface; structural level.}

\item{N_s}{the minimum sample size for each interval.}

\item{N_m}{the maximum sample size for each interval; default = 200.}
}
\value{
returns a list described above.
}
\description{
Generates simulated response for multilevel splines -- test function #2
}
\author{
YD Hwang \email{yhwang@g.skku.edu} and ER Lee \email{erlee@skku.edu}
}
