% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsplines_funs.R
\name{granular}
\alias{granular}
\title{Builds ``granular'' data}
\usage{
granular(x_k, y_k, mod)
}
\arguments{
\item{x_k}{design matrix}

\item{y_k}{response vector}

\item{mod}{underlying model; either `lm` or `glm`}
}
\description{
obtains the regression slope and its variance
certainly not optimal but this step shouldn't take long regardless
}
\author{
YD Hwang \email{yhwang@g.skku.edu} and ER Lee \email{erlee@skku.edu}
}
