% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsplines.R
\name{main_EM}
\alias{main_EM}
\title{Main EM function}
\usage{
main_EM(beta_hat_vec, V, K, lambda, maxit = 500)
}
\arguments{
\item{beta_hat_vec}{data vector of length J}

\item{V}{covariance matrix of size J by J}

\item{K}{kernel matrix from `make_K`}

\item{lambda}{tuning parameter}

\item{maxit}{maximum iteration number}
}
\description{
Running EM for multilevel splines
certainly not optimal...
}
\author{
YD Hwang \email{yhwang@g.skku.edu} and ER Lee \email{erlee@skku.edu}
}
