% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsplines.R
\name{make_K}
\alias{make_K}
\title{Generates kerel matrix}
\usage{
make_K(z)
}
\arguments{
\item{z}{Mid-interval value vector, it is safe to assume this to be equi-distant, but in principle it doesn't have to be. it's not tested though.}
}
\description{
Generates kernel matrix of J by J, where J = length(z) for multilevel splines
certainly not optimal but this step shouldn't take long regardless.
Used the formulation from Reinsch (1967).
}
\author{
YD Hwang \email{yhwang@g.skku.edu} and ER Lee \email{erlee@skku.edu}
}
