% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsplines.R
\docType{package}
\name{mlsplines}
\alias{mlsplines}
\alias{mlsplines-package}
\title{package for generate simulated response for multilevel splines}
\description{
Package for generating simulated response for multilevel splines, and fitting EM-based splines
}
\author{
YD Hwang \email{yhwang@g.skku.edu} and ER Lee \email{erlee@skku.edu}
}
