% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsplines.R
\name{naive_ss}
\alias{naive_ss}
\title{Naive strawman}
\usage{
naive_ss(beta_hat_vec, lambda, K)
}
\arguments{
\item{beta_hat_vec}{data vector of length J}

\item{lambda}{tuning parameter}

\item{K}{kernel matrix from `make_K`}
}
\description{
Running naive splines in \code{\link{mlsplines}}
}
\author{
YD Hwang \email{yhwang@g.skku.edu} and ER Lee \email{erlee@skku.edu}
}
