% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsplines_funs.R
\name{generate_response}
\alias{generate_response}
\title{generate simulated response for multilevel splines}
\usage{
generate_response(J, mod, e_sigma = 1, x_sigma = 1, z_sigma = 0.5,
  N_s, N_m = 200)
}
\arguments{
\item{J}{number of 'data' intervals}

\item{mod}{underlying model; either `lm` or `glm`}

\item{e_sigma}{error variance - around the mean function; data level.}

\item{x_sigma}{design matrix sigma}

\item{z_sigma}{error variance around my surface; structural level.}

\item{N_s}{the minimum sample size for each interval.}

\item{N_m}{the maximum sample size for each interval; default = 200.}
}
\description{
Generates simulated response for multilevel splines
}
\author{
YD Hwang \email{yhwang@g.skku.edu} and ER Lee \email{erlee@skku.edu}
}
