% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsplines.R
\name{make_K}
\alias{make_K}
\title{Generates kerel matrix}
\usage{
make_K(z)
}
\arguments{
\item{z}{mid-interval value vector. assumed the equi-distant}
}
\description{
Generates kernel matrix of J by J, where J = length(z) for multilevel splines in \code{\link{mlsplines}}
certainly not optimal but this step shouldn't take long regardless
}
\author{
YD Hwang \email{yhwang@g.skku.edu} and ER Lee \email{erlee@skku.edu}
}
